Convert altitudes and magnitudes to floats
Converted the altitude and magnitude values to floats. This is an
attempt to speed up computations on the gpu. Initial testing on my
machine (NVIDIA MX130) shows an average kernel speedup of the iterate
particle kernel from ~48ms to ~20ms. Setting up the simulation does
appear to take longer than before by about 1 second, this appears to
most likely come from the getLambdaAtS() function on line 83 of
dipoleB.cu. The increased error of using floats over doubles still needs
to be analyzed.

+++ include/BField/DipoleBLUT.h
@@ -20,8 +20,8 @@ protected:
 	#endif /* !__CUDA_ARCH__ */
 
 	//on device variables
-	double* altitude_d{ nullptr }; //serves as device-side C-style vector equivalent, and on host, pointer to this on dev
-	double* magnitude_d{ nullptr };
+	float* altitude_d{ nullptr }; //serves as device-side C-style vector equivalent, and on host, pointer to this on dev
+	float* magnitude_d{ nullptr };
 	
 	meters  simMin_m{ 0.0 };
 	meters  simMax_m{ 0.0 };
@@ -47,8 +47,8 @@ public:
 	__host__ __device__ double  getGradBAtS(const meters s, const seconds t) const override;
 	__host__ __device__ meters  getSAtAlt(const meters alt_fromRe) const override;
 
-    __device__          void    setAltArray(double* altArray);
-	__device__          void    setMagArray(double* magArray);
+    __device__          void    setAltArray(float* altArray);
+	__device__          void    setMagArray(float* magArray);
 
 	__host__            ratio   getErrTol() const;
 	__host__            meters  getds()     const;
