+++ include/physicalconstants.h
@@ -5,9 +5,9 @@
 constexpr double MASS_PROTON  { 1.672621898e-27 };	//kg
 constexpr double MASS_ELECTRON{ 9.10938356e-31 };	//kg
 constexpr double CHARGE_ELEM  { 1.6021766209e-19 }; //C
-constexpr double RADIUS_EARTH { 6.3712e6 };			//m
+constexpr float RADIUS_EARTH { 6.3712e6 };			//m
 constexpr double JOULE_PER_EV { 1.6021766209e-19 }; //J/eV
-constexpr double PI           { 3.14159265358979323846 };
-constexpr double RADS_PER_DEG { PI / 180.0 };       //rad/deg
+constexpr float PI           { 3.14159265358979323846 };
+constexpr float RADS_PER_DEG { PI / 180.0 };       //rad/deg
 
 #endif
